requirement.txt
click==7.1.2
Flask==1.1.2
Flask-SocketIO==4.3.0
imageio==2.8.0
imagezmq==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.4
opencv-python==4.2.0.34
Pillow==7.1.2
python-engineio==3.13.0
python-socketio==4.6.0
pyzmq==19.0.1
six==1.15.0
Werkzeug==1.0.1
